[MS] Moved input function to a separate file
